+++ src/components/SearchResults.tsx
@@ -6,12 +6,11 @@ import Grid from '@material-ui/core/Grid';
 import LifeEventPicker from './LifeEventPicker';
 import RelationshipPicker from './RelationshipPicker';
 import NamePicker from './NamePicker';
-// import SearchResultsTable from './SearchResultsTable';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import SearchResultsLocation from './SearchResultsLocations';
 import SearchResultsCategory from './SearchResultsCategory';
-import FixedTable from './FixedTable';
+import SearchResultsTable from './SearchResultsTable';
 
 export default function SearchResults(props: { setResults: any }): JSX.Element {
   const { setResults } = props;
@@ -45,8 +44,7 @@ export default function SearchResults(props: { setResults: any }): JSX.Element {
           <Typography component="h1" variant="h5">
             Search Results
           </Typography>
-          {/* <SearchResultsTable /> */}
-          <FixedTable />
+          <SearchResultsTable />
         </Grid>
       </Grid>
     </Container>
